// payload-types.ts
export interface Page {
  id: number;
  label: string;
  shownIn?: ('header' | 'footer')[] | null;
  slug: string;
  hero?: {
    title?: string | null;
    image?: (number | null) | Media;
    description?: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    impact?: ('high' | 'low') | null;
    enablePrimaryLink?: boolean | null;
    primaryLink?: {
      label?: string | null;
      path?: string | null;
    };
    enableSecondaryLink?: boolean | null;
    secondaryLink?: {
      label?: string | null;
      path?: string | null;
    };
  };
  blocks?:
    | (
        | CallToActionBlock
        | CardLinksBlock
        | CardListBlock
        | ColumnSectionBlock
        | ComparisonBlock
        | ContactFormBlock
        | DescriptionListBlock
        | EmbedMapBlock
        | FaqsBlock
        | FeaturedPostsBlock
        | GalleryBlock
        | LatestPostsBlock
        | LogoCloudBlock
        | MarqueeBlock
        | MediaBlock
        | PricingTableBlock
        | StatsBlock
        | TeamSectionBlock
        | TestimonialsBlock
      )[]
    | null;
  /**
   * Structured data for search engines and social platforms.
   */
  schemaMarkup?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
  };
  folder?: (number | null) | FolderInterface;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  humanReadableFilesize?: string | null;
  alt?: string | null;
  poster?: {
    relationTo: 'media';
    value: number | Media;
  } | null;
  caption?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    square?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    small?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    xlarge?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CallToActionBlock".
 */
export interface CallToActionBlock {
  title: string;
  description?: string | null;
  hasBackground?: boolean | null;
  button?: {
    label?: string | null;
    /**
     * It can start with "/" if it is inside the web or be a complete url (https://)
     */
    path?: string | null;
  };
  enableSecondaryButton?: boolean | null;
  secondaryButton?: {
    label?: string | null;
    path?: string | null;
  };
  id?: string | null;
  blockName?: string | null;
  blockType: 'callToAction';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CardLinksBlock".
 */
export interface CardLinksBlock {
  links?:
    | {
        title: string;
        label: string;
        /**
         * If it is an internal URL, it must start with "/".
         */
        url: string;
        /**
         * Images with vivid colors will work better. If not provided, the logo will be used.
         */
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'cardLinks';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "CardListBlock".
 */
export interface CardListBlock {
  items?:
    | {
        label: string;
        /**
         * Images with vivid colors will work better. If not provided, the logo will be used.
         */
        image?: (number | null) | Media;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'cardList';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ColumnSectionBlock".
 */
export interface ColumnSectionBlock {
  hasBackground?: boolean | null;
  columns?:
    | {
        size?: ('oneThird' | 'half' | 'twoThirds' | 'full') | null;
        content?: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        } | null;
        enableLink?: boolean | null;
        link?: {
          label?: string | null;
          url?: string | null;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'columnSection';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ComparisonBlock".
 */
export interface ComparisonBlock {
  title?: string | null;
  description?: string | null;
  beforeImage: number | Media;
  afterImage: number | Media;
  id?: string | null;
  blockName?: string | null;
  blockType: 'comparison';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContactFormBlock".
 */
export interface ContactFormBlock {
  title?: string | null;
  subtitle?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'contactForm';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "DescriptionListBlock".
 */
export interface DescriptionListBlock {
  title?: string | null;
  subtitle?: string | null;
  items?:
    | {
        title: string;
        content: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'descriptionList';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "EmbedMapBlock".
 */
export interface EmbedMapBlock {
  title?: string | null;
  description?: string | null;
  hasBackground?: boolean | null;
  googleMapsEmbedCode?: string | null;
  googleMapsUrl?: string | null;
  appleMapsUrl?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'embedMap';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FaqsBlock".
 */
export interface FaqsBlock {
  title?: string | null;
  subtitle?: string | null;
  items?:
    | {
        question: string;
        answer: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'faqs';
}
// ...
export interface Post {
  id: number;
  title: string;
  description: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  image?: (number | null) | Media;
  publishedAt?: string | null;
  authors?: (number | User)[] | null;
  featured?: boolean | null;
  slug: string;
  categories?: (number | Category)[] | null;
  relatedPosts?: (number | Post)[] | null;
  content: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  /**
   * Structured data for search engines and social platforms.
   */
  schemaMarkup?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
  };
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
  _status?: ('draft' | 'published') | null;
}
// ...
export interface HomePage {
  id: number;
  label: string;
  shownIn?: ('header' | 'footer')[] | null;
  hero?: {
    title?: string | null;
    image?: (number | null) | Media;
    description?: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    impact?: ('high' | 'low') | null;
    enablePrimaryLink?: boolean | null;
    primaryLink?: {
      label?: string | null;
      path?: string | null;
    };
    enableSecondaryLink?: boolean | null;
    secondaryLink?: {
      label?: string | null;
      path?: string | null;
    };
  };
  blocks?:
    | (
        | CallToActionBlock
        | CardLinksBlock
        | CardListBlock
        | ColumnSectionBlock
        | ComparisonBlock
        | ContactFormBlock
        | DescriptionListBlock
        | EmbedMapBlock
        | FaqsBlock
        | FeaturedPostsBlock
        | GalleryBlock
        | LatestPostsBlock
        | LogoCloudBlock
        | MarqueeBlock
        | MediaBlock
        | PricingTableBlock
        | StatsBlock
        | TeamSectionBlock
        | TestimonialsBlock
      )[]
    | null;
  /**
   * Structured data for search engines and social platforms.
   */
  schemaMarkup?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}